Optimisation: Image in memory
Thanks to this post http://csharpexamples.com/fast-image-processing-c/ the main algorithm now takes ~2 seconds instead of ~32 seconds.

The reason for that being because of IO operations hurting performance. As some may expect running the old version on Linux (I tried on Fedora 34) gave better results on a worse machine (~16 seconds).